github-upload-secrets: untangle secrets handling
Move all knowledge of how to deal with uploading secrets to the
(renamed) upload_secrets function.  This function now takes the
repository and environment, and a dictionary of secrets to upload, and
does all of the work, including path calculations, public key download,
(possible) environment creation, and uploading.

In order to accomplish this we detangle the way we deal with encryption
and now do it as the last step before uploading.  That allows some nice
cleanups elsewhere.