setup-deploy-keys: add cockpit-machines to node-cache
Setup the necessary deploy keys for cockpit-machines/ to use node-cache/
for its node_modules in the same way that cockpit/ does.

Note: we rename the COCKPIT_DEPLOY_KEY to SELF_DEPLOY_KEY in order to
reduce the differences between the workflows in each project.  We'll
backport that change to cockpit/ repo soon.