image-refresh: add explicit --propose argument
Instead of trying to do a heuristic about if we should propose the new
branch as a PR or not, use an explicit argument to control it.

In order to do that we need to move away from `task.main()`.  Drop
support for running from issues, while we're at it.

Finally, add support for posting the log URL, if we have it.